+++ src/ai_marketplace_monitor/ai.py
@@ -1,16 +1,16 @@
 import re
 import time
-from dataclasses import dataclass, field
+from dataclasses import asdict, dataclass, field
 from enum import Enum
 from logging import Logger
-from typing import Any, ClassVar, Generic, Type, TypeVar
+from typing import Any, ClassVar, Generic, Optional, Type, TypeVar
 
 from openai import OpenAI  # type: ignore
 from rich.pretty import pretty_repr
 
-from .item import SearchedItem
+from .listing import Listing
 from .marketplace import TItemConfig
-from .utils import CacheType, DataClassWithHandleFunc, cache, hilight
+from .utils import CacheType, CounterItem, DataClassWithHandleFunc, cache, counter, hilight
 
 
 class AIServiceProvider(Enum):
@@ -43,6 +43,24 @@ class AIResponse:
         if self.score > 3:
             return "succ"
         return "name"
+
+    @classmethod
+    def from_cache(
+        cls: Type["AIResponse"], listing: Listing, item_config: TItemConfig
+    ) -> Optional["AIResponse"]:
+        res = cache.get(
+            (CacheType.AI_INQUIRY.value, listing.marketplace, item_config.name, listing.id)
+        )
+        if res is None:
+            return None
+        return AIResponse(**res)
+
+    def to_cache(self: "AIResponse", listing: Listing, item_config: TItemConfig) -> None:
+        cache.set(
+            (CacheType.AI_INQUIRY.value, listing.marketplace, item_config.name, listing.id),
+            asdict(self),
+            tag=CacheType.AI_INQUIRY.value,
+        )
 
 
 @dataclass
@@ -123,7 +141,7 @@ class AIBackend(Generic[TAIConfig]):
     def connect(self: "AIBackend") -> None:
         raise NotImplementedError("Connect method must be implemented by subclasses.")
 
-    def get_prompt(self: "AIBackend", listing: SearchedItem, item_config: TItemConfig) -> str:
+    def get_prompt(self: "AIBackend", listing: Listing, item_config: TItemConfig) -> str:
         prompt = (
             f"""A user wants to buy a {item_config.name} from Facebook Marketplace. """
             f"""Search keywords: "{'" and "'.join(item_config.keywords)}", """
@@ -162,7 +180,7 @@ class AIBackend(Generic[TAIConfig]):
             self.logger.debug(f"""{hilight("[AI-Prompt]", "info")} {prompt}""")
         return prompt
 
-    def evaluate(self: "AIBackend", listing: SearchedItem, item_config: TItemConfig) -> AIResponse:
+    def evaluate(self: "AIBackend", listing: Listing, item_config: TItemConfig) -> AIResponse:
         raise NotImplementedError("Confirm method must be implemented by subclasses.")
 
 
@@ -185,20 +203,17 @@ class OpenAIBackend(AIBackend):
             if self.logger:
                 self.logger.info(f"""{hilight("[AI]", "name")} {self.config.name} connected.""")
 
-    def evaluate(
-        self: "OpenAIBackend", listing: SearchedItem, item_config: TItemConfig
-    ) -> AIResponse:
+    def evaluate(self: "OpenAIBackend", listing: Listing, item_config: TItemConfig) -> AIResponse:
         # ask openai to confirm the item is correct
+        counter.increment(CounterItem.AI_QUERY)
         prompt = self.get_prompt(listing, item_config)
-        cached_result = cache.get(
-            (CacheType.AI_INQUIRY.value, listing.marketplace, item_config.name, listing.id)
-        )
-        if cached_result is not None:
+        res: AIResponse | None = AIResponse.from_cache(listing, item_config)
+        if res is not None:
             if self.logger:
                 self.logger.info(
                     f"""{hilight("[AI]", "name")} {self.config.name} has already evaluated {hilight(listing.title)}."""
                 )
-            return AIResponse(cached_result["score"], cached_result["comment"])
+            return res
 
         self.connect()
 
@@ -239,6 +254,7 @@ class OpenAIBackend(AIBackend):
             or not answer.strip()
             or re.search(r"Rating[:\s]*[1-5]", answer, re.DOTALL) is None
         ):
+            counter.increment(CounterItem.FAILED_AI_QUERY)
             raise ValueError(f"Empty or invalid response from {self.config.name}: {response}")
 
         lines = answer.split("\n")
@@ -252,16 +268,13 @@ class OpenAIBackend(AIBackend):
                 comment = matched.group(2).strip()
                 break
 
-        cache.set(
-            (CacheType.AI_INQUIRY.value, listing.marketplace, item_config.name, listing.id),
-            {"score": score, "comment": comment},
-            tag=CacheType.AI_INQUIRY.value,
-        )
         res = AIResponse(score, comment)
+        res.to_cache(listing, item_config)
         if self.logger:
             self.logger.info(
                 f"""{hilight("[AI]", res.style)} {self.config.name} concludes {hilight(f"{res.conclusion} ({res.score}): {res.comment}", res.style)} for listing {hilight(listing.title)}."""
             )
+        counter.increment(CounterItem.NEW_AI_QUERY)
         return res