+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.6] - 2025-02-06
+
+- Allow sending reminders for available items after specified time. (#41)
+- Display counters
+
 ## [0.6.5] - 2025-02-05
 
 - Allow checking URLs during monitoring (#34)